+++ esp-hal/src/dma/mod.rs
@@ -1571,12 +1571,6 @@ pub trait DmaChannel: crate::private::Sealed + Sized {
 
     /// A description of the TX half of a DMA Channel.
     type Tx: TxRegisterAccess + InterruptAccess<DmaTxInterrupt>;
-
-    /// Returns the async interrupt handler.
-    fn async_handler<M: Mode>(ch: &Channel<'_, Self, M>) -> InterruptHandler;
-
-    /// Returns the interrupt.
-    fn interrupts<M: Mode>(ch: &Channel<'_, Self, M>) -> &'static [Interrupt];
 }
 
 #[doc(hidden)]
@@ -1666,16 +1660,17 @@ pub trait Rx: crate::private::Sealed {
 // DMA receive channel
 #[non_exhaustive]
 #[doc(hidden)]
-pub struct ChannelRx<'a, CH>
+pub struct ChannelRx<'a, CH, M>
 where
     CH: DmaChannel,
 {
     pub(crate) burst_mode: bool,
     pub(crate) rx_impl: CH::Rx,
+    pub(crate) mode: PhantomData<M>,
     pub(crate) _phantom: PhantomData<(&'a (), CH)>,
 }
 
-impl<'a, CH> ChannelRx<'a, CH>
+impl<'a, CH> ChannelRx<'a, CH, Blocking>
 where
     CH: DmaChannel,
 {
@@ -1688,19 +1683,74 @@ where
         Self {
             burst_mode: false,
             rx_impl,
+            mode: PhantomData,
             _phantom: PhantomData,
         }
     }
 
+    /// Converts a blocking channel to an async channel.
+    pub(crate) fn into_async(mut self) -> ChannelRx<'a, CH, Async> {
+        if let Some(handler) = self.rx_impl.async_handler() {
+            self.set_interrupt_handler(handler);
+        }
+        ChannelRx {
+            burst_mode: false,
+            rx_impl: self.rx_impl,
+            mode: PhantomData,
+            _phantom: PhantomData,
+        }
+    }
+
+    fn set_interrupt_handler(&mut self, handler: InterruptHandler)
+    where
+        CH: DmaChannel,
+    {
+        self.unlisten_in(EnumSet::all());
+        self.clear_in(EnumSet::all());
+
+        if let Some(interrupt) = self.rx_impl.peripheral_interrupt() {
+            for core in crate::Cpu::other() {
+                crate::interrupt::disable(core, interrupt);
+            }
+            unsafe { crate::interrupt::bind_interrupt(interrupt, handler.handler()) };
+            unwrap!(crate::interrupt::enable(interrupt, handler.priority()));
+        }
+    }
+}
+
+impl<'a, CH> ChannelRx<'a, CH, Async>
+where
+    CH: DmaChannel,
+{
+    /// Converts an async channel into a blocking channel.
+    pub(crate) fn into_blocking(self) -> ChannelRx<'a, CH, Blocking> {
+        if let Some(interrupt) = self.rx_impl.peripheral_interrupt() {
+            crate::interrupt::disable(Cpu::current(), interrupt);
+        }
+        ChannelRx {
+            burst_mode: false,
+            rx_impl: self.rx_impl,
+            mode: PhantomData,
+            _phantom: PhantomData,
+        }
+    }
+}
+
+impl<'a, CH, M> ChannelRx<'a, CH, M>
+where
+    CH: DmaChannel,
+    M: Mode,
+{
     /// Return a less specific (degraded) version of this channel.
     #[doc(hidden)]
-    pub fn degrade<DEG: DmaChannel>(self) -> ChannelRx<'a, DEG>
+    pub fn degrade<DEG: DmaChannel>(self) -> ChannelRx<'a, DEG, M>
     where
         CH: DmaChannelConvert<DEG>,
     {
         ChannelRx {
             burst_mode: self.burst_mode,
             rx_impl: CH::degrade_rx(self.rx_impl),
+            mode: PhantomData,
             _phantom: PhantomData,
         }
     }
@@ -1712,11 +1762,17 @@ where
     }
 }
 
-impl<CH> crate::private::Sealed for ChannelRx<'_, CH> where CH: DmaChannel {}
-
-impl<CH> Rx for ChannelRx<'_, CH>
+impl<CH, M> crate::private::Sealed for ChannelRx<'_, CH, M>
 where
     CH: DmaChannel,
+    M: Mode,
+{
+}
+
+impl<CH, M> Rx for ChannelRx<'_, CH, M>
+where
+    CH: DmaChannel,
+    M: Mode,
 {
     unsafe fn prepare_transfer_without_start(
         &mut self,
@@ -1894,17 +1950,18 @@ pub trait Tx: crate::private::Sealed {
 
 /// DMA transmit channel
 #[doc(hidden)]
-pub struct ChannelTx<'a, CH>
+pub struct ChannelTx<'a, CH, M>
 where
     CH: DmaChannel,
 {
     #[allow(unused)]
     pub(crate) burst_mode: bool,
     pub(crate) tx_impl: CH::Tx,
+    pub(crate) mode: PhantomData<M>,
     pub(crate) _phantom: PhantomData<(&'a (), CH)>,
 }
 
-impl<'a, CH> ChannelTx<'a, CH>
+impl<'a, CH> ChannelTx<'a, CH, Blocking>
 where
     CH: DmaChannel,
 {
@@ -1912,19 +1969,74 @@ where
         Self {
             burst_mode: false,
             tx_impl,
+            mode: PhantomData,
             _phantom: PhantomData,
         }
     }
 
+    /// Converts a blocking channel to an async channel.
+    pub(crate) fn into_async(mut self) -> ChannelTx<'a, CH, Async> {
+        if let Some(handler) = self.tx_impl.async_handler() {
+            self.set_interrupt_handler(handler);
+        }
+        ChannelTx {
+            burst_mode: false,
+            tx_impl: self.tx_impl,
+            mode: PhantomData,
+            _phantom: PhantomData,
+        }
+    }
+
+    fn set_interrupt_handler(&mut self, handler: InterruptHandler)
+    where
+        CH: DmaChannel,
+    {
+        self.unlisten_out(EnumSet::all());
+        self.clear_out(EnumSet::all());
+
+        if let Some(interrupt) = self.tx_impl.peripheral_interrupt() {
+            for core in crate::Cpu::other() {
+                crate::interrupt::disable(core, interrupt);
+            }
+            unsafe { crate::interrupt::bind_interrupt(interrupt, handler.handler()) };
+            unwrap!(crate::interrupt::enable(interrupt, handler.priority()));
+        }
+    }
+}
+
+impl<'a, CH> ChannelTx<'a, CH, Async>
+where
+    CH: DmaChannel,
+{
+    /// Converts an async channel into a blocking channel.
+    pub(crate) fn into_blocking(self) -> ChannelTx<'a, CH, Blocking> {
+        if let Some(interrupt) = self.tx_impl.peripheral_interrupt() {
+            crate::interrupt::disable(Cpu::current(), interrupt);
+        }
+        ChannelTx {
+            burst_mode: false,
+            tx_impl: self.tx_impl,
+            mode: PhantomData,
+            _phantom: PhantomData,
+        }
+    }
+}
+
+impl<'a, CH, M> ChannelTx<'a, CH, M>
+where
+    CH: DmaChannel,
+    M: Mode,
+{
     /// Return a less specific (degraded) version of this channel.
     #[doc(hidden)]
-    pub fn degrade<DEG: DmaChannel>(self) -> ChannelTx<'a, DEG>
+    pub fn degrade<DEG: DmaChannel>(self) -> ChannelTx<'a, DEG, M>
     where
         CH: DmaChannelConvert<DEG>,
     {
         ChannelTx {
             burst_mode: self.burst_mode,
             tx_impl: CH::degrade_tx(self.tx_impl),
+            mode: PhantomData,
             _phantom: PhantomData,
         }
     }
@@ -1936,11 +2048,17 @@ where
     }
 }
 
-impl<CH> crate::private::Sealed for ChannelTx<'_, CH> where CH: DmaChannel {}
-
-impl<CH> Tx for ChannelTx<'_, CH>
+impl<CH, M> crate::private::Sealed for ChannelTx<'_, CH, M>
 where
     CH: DmaChannel,
+    M: Mode,
+{
+}
+
+impl<CH, M> Tx for ChannelTx<'_, CH, M>
+where
+    CH: DmaChannel,
+    M: Mode,
 {
     unsafe fn prepare_transfer_without_start(
         &mut self,
@@ -2116,6 +2234,9 @@ pub trait RegisterAccess: crate::private::Sealed {
 pub trait RxRegisterAccess: RegisterAccess {
     #[cfg(gdma)]
     fn set_mem2mem_mode(&self, value: bool);
+
+    fn peripheral_interrupt(&self) -> Option<Interrupt>;
+    fn async_handler(&self) -> Option<InterruptHandler>;
 }
 
 #[doc(hidden)]
@@ -2125,6 +2246,9 @@ pub trait TxRegisterAccess: RegisterAccess {
 
     /// Outlink descriptor address when EOF occurs of Tx channel.
     fn last_dscr_address(&self) -> usize;
+
+    fn peripheral_interrupt(&self) -> Option<Interrupt>;
+    fn async_handler(&self) -> Option<InterruptHandler>;
 }
 
 #[doc(hidden)]
@@ -2154,10 +2278,9 @@ where
     M: Mode,
 {
     /// RX half of the channel
-    pub rx: ChannelRx<'d, CH>,
+    pub rx: ChannelRx<'d, CH, M>,
     /// TX half of the channel
-    pub tx: ChannelTx<'d, CH>,
-    pub(crate) phantom: PhantomData<M>,
+    pub tx: ChannelTx<'d, CH, M>,
 }
 
 impl<'d, C> Channel<'d, C, Blocking>
@@ -2171,15 +2294,8 @@ where
     where
         C: DmaChannel,
     {
-        self.unlisten(EnumSet::all());
-        self.clear_interrupts(EnumSet::all());
-        for interrupt in C::interrupts(self).iter().copied() {
-            for core in crate::Cpu::other() {
-                crate::interrupt::disable(core, interrupt);
-            }
-            unsafe { crate::interrupt::bind_interrupt(interrupt, handler.handler()) };
-            unwrap!(crate::interrupt::enable(interrupt, handler.priority()));
-        }
+        self.rx.set_interrupt_handler(handler);
+        self.tx.set_interrupt_handler(handler);
     }
 
     /// Listen for the given interrupts
@@ -2226,18 +2342,15 @@ where
 
     /// Configure the channel.
     pub fn configure(&mut self, burst_mode: bool, priority: DmaPriority) {
-        self.tx.configure(burst_mode, priority);
         self.rx.configure(burst_mode, priority);
+        self.tx.configure(burst_mode, priority);
     }
 
     /// Converts a blocking channel to an async channel.
-    pub fn into_async(mut self) -> Channel<'d, C, Async> {
-        self.set_interrupt_handler(C::async_handler(&self));
-
+    pub fn into_async(self) -> Channel<'d, C, Async> {
         Channel {
-            tx: self.tx,
-            rx: self.rx,
-            phantom: PhantomData,
+            rx: self.rx.into_async(),
+            tx: self.tx.into_async(),
         }
     }
 }
@@ -2248,14 +2361,9 @@ where
 {
     /// Converts an async channel to a blocking channel.
     pub fn into_blocking(self) -> Channel<'d, C, Blocking> {
-        for interrupt in C::interrupts(&self).iter().copied() {
-            crate::interrupt::disable(Cpu::current(), interrupt);
-        }
-
         Channel {
-            tx: self.tx,
-            rx: self.rx,
-            phantom: PhantomData,
+            rx: self.rx.into_blocking(),
+            tx: self.tx.into_blocking(),
         }
     }
 }
@@ -2286,7 +2394,6 @@ where
         Channel {
             rx: self.rx.degrade(),
             tx: self.tx.degrade(),
-            phantom: PhantomData,
         }
     }
 }