+++ esp-hal/src/parl_io.rs
@@ -811,7 +811,7 @@ pub struct ParlIoTx<'d, DM>
 where
     DM: Mode,
 {
-    tx_channel: ChannelTx<'d, <PARL_IO as DmaEligible>::Dma>,
+    tx_channel: ChannelTx<'d, <PARL_IO as DmaEligible>::Dma, DM>,
     tx_chain: DescriptorChain,
     phantom: PhantomData<DM>,
 }
@@ -890,7 +890,7 @@ pub struct ParlIoRx<'d, DM>
 where
     DM: Mode,
 {
-    rx_channel: ChannelRx<'d, <PARL_IO as DmaEligible>::Dma>,
+    rx_channel: ChannelRx<'d, <PARL_IO as DmaEligible>::Dma, DM>,
     rx_chain: DescriptorChain,
     phantom: PhantomData<DM>,
 }
@@ -1034,20 +1034,14 @@ impl<'d> ParlIoFullDuplex<'d, Blocking> {
 
     /// Convert to an async version.
     pub fn into_async(self) -> ParlIoFullDuplex<'d, Async> {
-        let channel = Channel {
-            tx: self.tx.tx_channel,
-            rx: self.rx.rx_channel,
-            phantom: PhantomData::<Blocking>,
-        };
-        let channel = channel.into_async();
         ParlIoFullDuplex {
             tx: TxCreatorFullDuplex {
-                tx_channel: channel.tx,
+                tx_channel: self.tx.tx_channel.into_async(),
                 descriptors: self.tx.descriptors,
                 phantom: PhantomData,
             },
             rx: RxCreatorFullDuplex {
-                rx_channel: channel.rx,
+                rx_channel: self.rx.rx_channel.into_async(),
                 descriptors: self.rx.descriptors,
                 phantom: PhantomData,
             },
@@ -1094,20 +1088,14 @@ impl InterruptConfigurable for ParlIoFullDuplex<'_, Blocking> {
 impl<'d> ParlIoFullDuplex<'d, Async> {
     /// Convert to a blocking version.
     pub fn into_blocking(self) -> ParlIoFullDuplex<'d, Blocking> {
-        let channel = Channel {
-            tx: self.tx.tx_channel,
-            rx: self.rx.rx_channel,
-            phantom: PhantomData::<Async>,
-        };
-        let channel = channel.into_blocking();
         ParlIoFullDuplex {
             tx: TxCreatorFullDuplex {
-                tx_channel: channel.tx,
+                tx_channel: self.tx.tx_channel.into_blocking(),
                 descriptors: self.tx.descriptors,
                 phantom: PhantomData,
             },
             rx: RxCreatorFullDuplex {
-                rx_channel: channel.rx,
+                rx_channel: self.rx.rx_channel.into_blocking(),
                 descriptors: self.rx.descriptors,
                 phantom: PhantomData,
             },
@@ -1372,7 +1360,7 @@ impl<'d, DM> DmaSupportTx for ParlIoTx<'d, DM>
 where
     DM: Mode,
 {
-    type TX = ChannelTx<'d, <PARL_IO as DmaEligible>::Dma>;
+    type TX = ChannelTx<'d, <PARL_IO as DmaEligible>::Dma, DM>;
 
     fn tx(&mut self) -> &mut Self::TX {
         &mut self.tx_channel
@@ -1414,7 +1402,7 @@ where
     }
 
     fn start_receive_bytes_dma(
-        rx_channel: &mut ChannelRx<'d, <PARL_IO as DmaEligible>::Dma>,
+        rx_channel: &mut ChannelRx<'d, <PARL_IO as DmaEligible>::Dma, DM>,
         rx_chain: &mut DescriptorChain,
         ptr: *mut u8,
         len: usize,
@@ -1468,7 +1456,7 @@ impl<'d, DM> DmaSupportRx for ParlIoRx<'d, DM>
 where
     DM: Mode,
 {
-    type RX = ChannelRx<'d, <PARL_IO as DmaEligible>::Dma>;
+    type RX = ChannelRx<'d, <PARL_IO as DmaEligible>::Dma, DM>;
 
     fn rx(&mut self) -> &mut Self::RX {
         &mut self.rx_channel
@@ -1484,7 +1472,7 @@ pub struct TxCreator<'d, DM>
 where
     DM: Mode,
 {
-    tx_channel: ChannelTx<'d, <PARL_IO as DmaEligible>::Dma>,
+    tx_channel: ChannelTx<'d, <PARL_IO as DmaEligible>::Dma, DM>,
     descriptors: &'static mut [DmaDescriptor],
     phantom: PhantomData<DM>,
 }
@@ -1494,7 +1482,7 @@ pub struct RxCreator<'d, DM>
 where
     DM: Mode,
 {
-    rx_channel: ChannelRx<'d, <PARL_IO as DmaEligible>::Dma>,
+    rx_channel: ChannelRx<'d, <PARL_IO as DmaEligible>::Dma, DM>,
     descriptors: &'static mut [DmaDescriptor],
     phantom: PhantomData<DM>,
 }
@@ -1504,7 +1492,7 @@ pub struct TxCreatorFullDuplex<'d, DM>
 where
     DM: Mode,
 {
-    tx_channel: ChannelTx<'d, <PARL_IO as DmaEligible>::Dma>,
+    tx_channel: ChannelTx<'d, <PARL_IO as DmaEligible>::Dma, DM>,
     descriptors: &'static mut [DmaDescriptor],
     phantom: PhantomData<DM>,
 }
@@ -1514,7 +1502,7 @@ pub struct RxCreatorFullDuplex<'d, DM>
 where
     DM: Mode,
 {
-    rx_channel: ChannelRx<'d, <PARL_IO as DmaEligible>::Dma>,
+    rx_channel: ChannelRx<'d, <PARL_IO as DmaEligible>::Dma, DM>,
     descriptors: &'static mut [DmaDescriptor],
     phantom: PhantomData<DM>,
 }